+++ backends-velox/src/test/scala/io/glutenproject/execution/TestOperator.scala
@@ -20,7 +20,7 @@ import io.glutenproject.GlutenConfig
 import io.glutenproject.sql.shims.SparkShimLoader
 
 import org.apache.spark.SparkConf
-import org.apache.spark.sql.{AnalysisException, DataFrame, Row}
+import org.apache.spark.sql.{AnalysisException, Row}
 import org.apache.spark.sql.execution.{FilterExec, GenerateExec, ProjectExec, RDDScanExec}
 import org.apache.spark.sql.functions.{avg, col, lit, udf}
 import org.apache.spark.sql.internal.SQLConf
@@ -294,85 +294,6 @@ class TestOperator extends VeloxWholeStageTransformerSuite {
     }
   }
 
-  test("chr function") {
-    val df = runQueryAndCompare(
-      "SELECT chr(l_orderkey + 64) " +
-        "from lineitem limit 1") { _ => }
-    checkLengthAndPlan(df, 1)
-  }
-
-  test("bin function") {
-    val df = runQueryAndCompare(
-      "SELECT bin(l_orderkey) " +
-        "from lineitem limit 1") {
-      checkOperatorMatch[ProjectExecTransformer]
-    }
-    checkLengthAndPlan(df, 1)
-  }
-
-  test("abs function") {
-    val df = runQueryAndCompare(
-      "SELECT abs(l_orderkey) " +
-        "from lineitem limit 1") { _ => }
-    checkLengthAndPlan(df, 1)
-  }
-
-  test("ceil function") {
-    val df = runQueryAndCompare(
-      "SELECT ceil(cast(l_orderkey as long)) " +
-        "from lineitem limit 1") { _ => }
-    checkLengthAndPlan(df, 1)
-  }
-
-  test("floor function") {
-    val df = runQueryAndCompare(
-      "SELECT floor(cast(l_orderkey as long)) " +
-        "from lineitem limit 1") { _ => }
-    checkLengthAndPlan(df, 1)
-  }
-
-  test("exp function") {
-    val df = spark.sql("SELECT exp(l_orderkey) from lineitem limit 1")
-    checkLengthAndPlan(df, 1)
-  }
-
-  test("power function") {
-    val df = runQueryAndCompare(
-      "SELECT power(l_orderkey, 2.0) " +
-        "from lineitem limit 1") { _ => }
-    checkLengthAndPlan(df, 1)
-  }
-
-  test("pmod function") {
-    val df = runQueryAndCompare(
-      "SELECT pmod(cast(l_orderkey as int), 3) " +
-        "from lineitem limit 1") { _ => }
-    checkLengthAndPlan(df, 1)
-  }
-
-  test("round function") {
-    val df = runQueryAndCompare(
-      "SELECT round(cast(l_orderkey as int), 2)" +
-        "from lineitem limit 1")(checkOperatorMatch[ProjectExecTransformer])
-  }
-
-  test("greatest function") {
-    val df = runQueryAndCompare(
-      "SELECT greatest(l_orderkey, l_orderkey)" +
-        "from lineitem limit 1")(checkOperatorMatch[ProjectExecTransformer])
-  }
-
-  test("least function") {
-    val df = runQueryAndCompare(
-      "SELECT least(l_orderkey, l_orderkey)" +
-        "from lineitem limit 1")(checkOperatorMatch[ProjectExecTransformer])
-  }
-
-  // Test "SELECT ..." without a from clause.
-  test("isnull function") {
-    runQueryAndCompare("SELECT isnull(1)") { _ => }
-  }
-
   test("df.count()") {
     val df = runQueryAndCompare("select * from lineitem limit 1") { _ => }
     checkLengthAndPlan(df, 1)
@@ -432,14 +353,6 @@ class TestOperator extends VeloxWholeStageTransformerSuite {
                          |) where l_suppkey != 0 limit 100;
                          |""".stripMargin) {
       checkOperatorMatch[LimitTransformer]
-    }
-  }
-
-  test("round") {
-    runQueryAndCompare("""
-                         |select round(l_quantity, 2) from lineitem;
-                         |""".stripMargin) {
-      checkOperatorMatch[ProjectExecTransformer]
     }
   }
 
@@ -1118,14 +1031,6 @@ class TestOperator extends VeloxWholeStageTransformerSuite {
       spark.sql("INSERT INTO TABLE remainder VALUES(0, null)")
 
       runQueryAndCompare("SELECT c1 % c2 FROM remainder")(df => checkFallbackOperators(df, 0))
-    }
-  }
-
-  test("Support HOUR function") {
-    withTable("t1") {
-      sql("create table t1 (c1 int, c2 timestamp) USING PARQUET")
-      sql("INSERT INTO t1 VALUES(1, NOW())")
-      runQueryAndCompare("SELECT c1, HOUR(c2) FROM t1 LIMIT 1")(df => checkFallbackOperators(df, 0))
     }
   }
